% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_dual.R
\name{Check_dual}
\alias{Check_dual}
\title{Quality Control for Post-processing}
\usage{
Check_dual(data, par, res, g, lambda = c(8, 4))
}
\arguments{
\item{data}{The CNV dataset prepared for lasso (see \code{\link[graphics]{data_lasso}})}

\item{par}{The parameter prepared for lasso (see \code{\link[graphics]{par_lasso}})}

\item{res}{The result after solving lasso problem  (see \code{\link[graphics]{Run_lasso}})}

\item{g}{Integer-valued ploidy}

\item{lambda}{L1-Penalty for all the variables (see \code{\link[graphics]{Run_lasso}})}
}
\value{
None
}
\description{
Check the dual problem of the lasso problem helps us to know how far we are from optimal.
}
\details{
This function would print out 5 values.

The first 2 are the dual variable and the 3rd and 4th should be your lambda parameters, which were used for \code{\link[graphics]{Run_lasso}}.
To achieve optimality, we should have the first 2 smaller or equal to the 3rd and 4th respectively.
However, due to numeric error, sometimes it is slightly larger (less than 0.01%), which is also acceptable.

The last one is a duality gap, but it is a valid value only when the requirement above is satisfied.
And it is strictly positive when it is valid.
However, it should be a small value in most cases.
}
\examples{

Check_dual(wkdata,par,Lasso_res,g_int)

}
