% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lasso.R
\name{data_lasso}
\alias{data_lasso}
\title{Prepare Dataset for Post-processing}
\usage{
data_lasso(data, g, par_a = 1)
}
\arguments{
\item{data}{The CNV dataset (see \code{\link[graphics]{load_data}})}

\item{g}{The integer-valued ploidy}

\item{par_a}{The constant indicating how ploidy infecting the final copy number profile}
}
\value{
List of values or matrice for lasso problem
}
\description{
Modify the original dataset in order to save some steps in the approximal gradient descent iteration for lasso.
}
\examples{

wkdata<-data_lasso(wkdata,g_int)

}
