% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_sample.R
\name{down_sample}
\alias{down_sample}
\title{Down Sample The Original Dataset}
\usage{
down_sample(CNV_data, factor = 10)
}
\arguments{
\item{CNV_data}{The CNV dataset (see \code{\link[graphics]{load_data}})}

\item{factor}{The factore to which the dataset to do down sampled}
}
\value{
Dataset with the minor and major copy, as well as the number of loci being down sampled (see \code{\link[graphics]{load_data}})
}
\description{
Down-sample the dataset in order to run faster.
}
\details{
For the copy number dataset is usually quite large to work with, and the fixed-point iteration is quite computationally expensive,
to save some time, it is helpful to down-sample the number of loci (no change to the number of samples).

This function down-samples the original one to make the algorithms run faster, and because most loci have similar values to the adjacent one,
this doesn't cause much inaccuracy.
}
\examples{

wkdata<-down_sample(wkdata,10)

}
\seealso{
\code{\link{load_data}}
}
