% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPI.R
\name{FPI}
\alias{FPI}
\title{Fixed Point Iteration}
\usage{
FPI(data, par, var, new, max_iter = 400)
}
\arguments{
\item{data}{The CNV dataset (see \code{\link[graphics]{load_data}})}

\item{par}{The parameter used for fixed point method. (see \code{\link[graphics]{set_par}})}

\item{var}{The current value for all variables (see \code{\link[graphics]{init_var}})}

\item{new}{The initial loss value}

\item{max_iter}{The maximum number of iterations that the method is allowed to run}
}
\value{
list of the optimal values for variables and the loss for each iteration
}
\description{
Run fixed point iteration.
}
\details{
This is the main part for WGD analysis. 
This function uses the fixed point method to iteratively calculate the optimal value for all variables and store the loss for each iteration in an array.

It runs until the improvement in loss is sufficiently small or the number of iteration reaches the max defined by user.
It then returns a list, containing two objects: var, which is a list for the optimal value of the variables; loss_array, which is an array of loss values.
}
\examples{

FPI_result=FPI(wkdata,CNV_par,common_var,init_loss)

}
