% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sep.R
\name{get_sep}
\alias{get_sep}
\title{Calculate the Separation Value for Ploidy}
\usage{
get_sep(var)
}
\arguments{
\item{var}{The current value for all variables (see \code{\link[graphics]{init_var}})}
}
\value{
Value for the separation of ploidy
}
\description{
In order to make the results more biologically meaningful, we want to set the (haplotype) ploidy to either 1 or 2, 
indicating whether such sample has genome duplication or not.
This function finds the valley for the distribution of inferref ploidy from FPI, and then we could separate the ploidy based on this value (see \code{\link[graphics]{int_ploidy}}).
}
\examples{

g_sep=get_sep(var)

}
