% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_var.R
\name{init_var}
\alias{init_var}
\title{Set the Initial Value for Variables}
\usage{
init_var(data, par)
}
\arguments{
\item{data}{The CNV dataset (see \code{\link[graphics]{load_data}})}

\item{par}{The parameter used for fixed point method. (see \code{\link[graphics]{set_par}})}
}
\value{
List of three matrice containing initial values for ploidy and recurrent CNV before and after WGD.
}
\description{
In order to run the fixed-point method, we need to start from some initial value.
This function set the initial values for all the variables to their mean values.
}
\examples{

common_var<-init_var(wkdata,CNV_par)

}
