% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_ploidy.R
\name{int_ploidy}
\alias{int_ploidy}
\title{Calculate Integer Ploidy Value for Post-processing}
\usage{
int_ploidy(var, sep)
}
\arguments{
\item{var}{The current value for all variables (see \code{\link[graphics]{init_var}})}

\item{sep}{The separation value (see \code{\link[graphics]{get_sep}})}
}
\value{
Cector containing integer-valued ploidy
}
\description{
Set (haplotype) ploidy value to either 1 or 2.
}
\details{
To get the interger ploidy value, 
we set all the infered ploidy value from FPI to be either 1 or 2, 
based on their value relative to the seperation value.

This gives us (haplotype) ploidy values for all samples, based on which we could run a lasso-like post-processing step.
The post-processing could make the prediction for recurrent CNV before and after WGD more biologically meaningful.
}
\examples{

g_int=int_ploidy(FPI_result$var,g_sep)

}
