% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Dataset Containing Copy Number Profile}
\usage{
load_data(mainDir, cancer_type, coor = "coords.txt")
}
\arguments{
\item{mainDir}{The directory which contains all the files.}

\item{cancer_type}{The type of the cancer: the files are named as "cancer_type"_minorcn.csv or "cancer_type"_majorcn.csv}

\item{coor}{The name for the coordinate file. Default as coords.txt.}
}
\value{
A list containing: the number of samples, number of loci, all minor copies, all major copies, all loci
}
\description{
Read text files containing copy number profiles called for multiple samples.
}
\details{
There are three different csv files containing minor copy, major copy and coordinates.

This returns list containing all the input data we need for down-stream analysis.
The list consists of two values, which indicate the number of samples and loci respectively,
and three data frame which contain minor copy, major copy, and coordinates.
}
\examples{

wkdata<-load_data("C:\\\\Double","BLCA")

}
