% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_after.R
\name{plot_after}
\alias{plot_after}
\title{Plot Recurrent CNV After WGD for each loci}
\usage{
plot_after(data, var)
}
\arguments{
\item{data}{the CNV dataset (see \code{\link[graphics]{load_data}})}

\item{var}{The optimal value for all variables (see \code{\link[graphics]{Run_lasso}})}
}
\value{
ggplot2 graph object
}
\description{
Plot the values for the common events happenning after WGD for each loci
ggplot2 is needed for this function
}
\examples{

library(ggplot2)
p_after<-plot_after(wkdata,Lasso_res$var)

}
