% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot the Original Copy number for Selected Loci}
\usage{
plot_bar(data, g, loci, max)
}
\arguments{
\item{data}{The CNV dataset (see \code{\link[graphics]{load_data}})}

\item{g}{Inferred ploidy value (see \code{\link[graphics]{}})}

\item{loci}{The particular loci to be emphasized}

\item{max}{The max number of copies for all loci}
}
\value{
ggplot2 graph object
}
\description{
To study some particular loci, this function count the number of samples 
containing certain number of loci in samples with and without WGD
}
\examples{

library(ggplot2)
p_bar<-plot_bar(wkdata,g_int,loci=c(3606,864,4858,5261,5570),max=7)

}
