% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_before.R
\name{plot_before}
\alias{plot_before}
\title{Plot Recurrent CNV Before WGD for each loci}
\usage{
plot_before(data, var)
}
\arguments{
\item{data}{the CNV dataset (see \code{\link[graphics]{load_data}})}

\item{var}{The optimal value for all variables (see \code{\link[graphics]{Run_lasso}})}
}
\value{
ggplot2 graph object
}
\description{
Plot the values for the common events happenning before WGD for each loci
ggplot2 is needed for this function
}
\examples{

library(ggplot2)
p_before<-plot_before(wkdata,Lasso_res$var)

}
