% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ploidy.R
\name{plot_ploidy}
\alias{plot_ploidy}
\title{Plot Inferred v.s. Average Ploidy}
\usage{
plot_ploidy(data, var, g_sep)
}
\arguments{
\item{data}{the CNV dataset (see \code{\link[graphics]{load_data}})}

\item{var}{The optimal value for all variables (see \code{\link[graphics]{init_var}})}

\item{g_sep}{The separation value for inferred ploidy}
}
\value{
ggplot2 graph object
}
\description{
Plot the inferred ploidy from FPI to average ploidy.
ggplot2 is needed for this function
}
\examples{

library(ggplot2)
p<-plot_ploidy(wkdata,FPI_result$var,g_sep)

}
