% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_par.R
\name{set_par}
\alias{set_par}
\title{Set Up Parameters for Analysis}
\usage{
set_par(data, r_CNV = c(10, 10), r_ploidy = 5, r_discontinuity = 0.5,
  par_a_m = 1, par_a_p = 1, A_m = c(1, 1), B_m = c(0, -1),
  Mean_m = c(0, -1), A_p = c(1, 1), B_p = c(0, -1), Mean_p = c(0, 0))
}
\arguments{
\item{data}{The CNV dataset (see \code{\link[graphics]{load_data}})}

\item{r_CNV}{Regularizer penalty for all recurrent CNV}

\item{r_ploidy}{Regularizer penalty for (haplotype) ploiidy}

\item{r_discontinuity}{Regularizer penalty for the discotinuity between two adjacent loci}

\item{par_a_m, par_a_p}{The constant indicating how ploidy infecting the final copy number profile for minor and major copy respectively}

\item{A_m, A_p}{A vector containing the coeffient of the effect of the recurrent CNV with respect to the ploidy value for minor and major copy respectively}

\item{B_m, B_p}{A vector containing the coeffient of the effect of the recurrent CNV with respect to the constant value for minor and major copy respectively}

\item{Mean_m, Mean_p}{A vector containing mean values of all the recurrent CNV for minor and major copy respectively}
}
\value{
List of parameters used for analysis
}
\description{
Set up the parameters for the fixed point method.
}
\details{
The 2nd and 3rd input are the regularizers for the L2 penalty of CNV before and after WGD, and (haplotype) ploidy from their mean.
The 4th input r_discontinuity is a penalty for the difference between two adjacent loci. If this is set to 0, then identity matrix would be used as matrix L in the model.

The following parameters indicate the types of recurrent CNV that should be included into the model.
The default setting is the changes before WGD affecting all samples and changes after WGD affecting only samples with WGD.
However, for changes after WGD, the mean value is set to 0 for major copy, but -1 to minor copy, because it is common to see 1 copy number loss after WGD on minor copy.
A_m,B_m and Mean_m should have the same lenth and so do the ones for major copy. And they should all be smaller or equal to the length of r_CNV.

This function will return all these parameters in a list for fixed-ponit method.
}
\examples{

CNV_par<-set_par(wkdata)

}
