% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_lasso.R
\name{var_lasso}
\alias{var_lasso}
\title{Prepare Variables for Post-processing}
\usage{
var_lasso(var, factor = 10)
}
\arguments{
\item{var}{The value for all variables from FPI (see \code{\link[graphics]{init_var}})}

\item{factor}{The factor to which you want to up-sample the variable from FPI (should be the same as the down-sample factor used before)}
}
\value{
matrice containing the recurrent CNV before and after WGD
}
\description{
Set up initial values for variables used for lasso problem.
}
\details{
Since in most cases, fixed-point method is run before post-processing. 
It saves a bit time if we start from that result. 
We can just replicate that result by the factor which was used before during down-sampling, to get an initial value for post-processing.
}
\examples{

var<-var_lasso(FPI_result$var)

}
